# backend/docs/v2/v2_swagger.yaml
basePath: /api/v2
definitions:
...
  CreatePortfolioV2Request:
    properties:
      base_currency:
        type: string
      benchmark:
        type: string
      code:
        maxLength: 40
        type: string
      description:
        type: string
      fund_code:
        maxLength: 40
        type: string
      has_units:
        type: boolean
      inception_date:
        type: string
      manager_user_id:
        type: string
      name:
        maxLength: 255
        type: string
      portfolio_type:
        enum:
        - LIVE
        - SIMULATION
        - MODEL
        type: string
      risk_profile:
        type: string
      strategy_code:
        type: string
      style_id:
        type: string
      tax_lot_method:
        type: string
      valuation_currency:
        type: string
    required:
    - base_currency
    - code
    - fund_code
    - inception_date
    - name
    - portfolio_type
    - valuation_currency
    type: object
...
  ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      error:
        type: string
    type: object
...
  PortfolioListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/PortfolioResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Portfolio V2 endpoints (docs/api/portfolio-v2-api-ddd.md), mounted
    under /api/v2.
  title: IMS Thailand API — Portfolio V2
  version: 2.0.0
paths:
  /portfolios:
    get:
      description: List portfolios visible to the authenticated user, scoped to their
        accessible funds (Portfolio V2).
      parameters:
      - description: Fund code
        in: query
        name: fund_code
        type: string
      - description: Portfolio status
        in: query
        name: status
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 50, max 200)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PortfolioListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: List Portfolios (V2)
      tags:
      - Investment - Portfolios V2
    post:
      consumes:
      - application/json
      description: Create a portfolio under a fund, resolved by business fund_code
        (Portfolio V2). Request body must not include fund_id.
      parameters:
      - description: Portfolio create payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreatePortfolioV2Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/PortfolioResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Portfolio (V2)
      tags:
      - Investment - Portfolios V2
  /portfolios/{portfolioCode}:
    get:
      description: Retrieve one portfolio by its business code (Portfolio V2).
      parameters:
      - description: Portfolio code
        in: path
        name: portfolioCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PortfolioResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Portfolio By Code
      tags:
      - Investment - Portfolios V2
    patch:
      consumes:
      - application/json
      description: Update mutable descriptive metadata on a portfolio, resolved by
        business code (Portfolio V2), using optimistic version control. Never changes
        fund association, code, or lifecycle status.
      parameters:
      - description: Portfolio code
        in: path
        name: portfolioCode
        required: true
        type: string
      - description: Portfolio patch payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PatchPortfolioV2Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PortfolioResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Portfolio By Code
      tags:
      - Investment - Portfolios V2
  /portfolios/{portfolioCode}/cash:
    get:
      description: List cash balances by currency for a portfolio, resolved by business
        code (Portfolio V2).
      parameters:
      - description: Portfolio code
        in: path
        name: portfolioCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/CashBalanceResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Portfolio Cash Balances By Code
      tags:
      - Investment - Portfolios V2
...
  /portfolios/{portfolioCode}/holdings:
    get:
      description: List current holdings for a portfolio, resolved by business code
        (Portfolio V2).
      parameters:
      - description: Portfolio code
        in: path
        name: portfolioCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/HoldingResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Portfolio Holdings By Code
      tags:
      - Investment - Portfolios V2
...
  /portfolios/{portfolioCode}/valuations:
    get:
      description: List valuation snapshots for a portfolio, resolved by business
        code (Portfolio V2).
      parameters:
      - description: Portfolio code
        in: path
        name: portfolioCode
        required: true
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 50, max 200)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ValuationListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: List Portfolio Valuations By Code
      tags:
      - Investment - Portfolios V2
  /portfolios/{portfolioCode}/valuations/latest:
    get:
      description: Retrieve the latest internal valuation snapshot for a portfolio,
        resolved by business code (Portfolio V2).
      parameters:
      - description: Portfolio code
        in: path
        name: portfolioCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ValuationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Latest Portfolio Valuation By Code
      tags:
      - Investment - Portfolios V2
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer Token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"